fix(test): replace "pidof sshd" for /run/sshd.pid
In test_balloon.py we are trying to find the PID of the SSH daemon
process, so that we later change its OOM score, so that it does not get
killed when we (deliberately) exhaust the available memory in a microVM.
We use "pidof sshd" to do that. However, pidof returns the PIDs of all
threads of the process. Now sshd launches a new thread for every new SSH
connection. Later when we iterate over these PIDs to change their OOM
score not all of the threads might be there, so the choom will fail for
these. This is not really a problem, but it some times leads to
misleading error messages.

This commit drops the use of "pidof" in favour of reading the daemon's
PID from "/run/sshd.pid".

Signed-off-by: Babis Chalios <bchalios@amazon.es>