test: disable pylint's redefined-outer-name
This lint unilaterally forbids shadowing. However, this clashes with
pytest fixtures in a great many places: To use a fixture a test function
needs to have an argument that matches the fixture's name. However, if
the fixture is defined in the same file as its usage, then this will
trigger pylint's redfined-outer-name. Since this is a very common
pattern, disable this lint.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>