Remove redundant min-version check from artifacts.py
With the new CI artifacts system, we no longer keep all firecracker
releases in a single S3 bucket, and instead keep a folder of supported
firecracker releases per artifacts folder. This means that in the 1.6
artifacts folder, only 1.4 and 1.5 release binaries exist. Therefore,
the additional check in firecracker_artifacts() is no longer needed. We
remove it, because it specifies the minimal version as a tuple, and is
thus easily missed when doign a release (as the runbook only hints that
sometimes min-versions are specified as strings).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>