feat(metrics)(net): Use iface_id to emit per net device metrics
With Vec based approach, "net0" is metrics for the 1st net device
created. However, if devices are not created in the same order all
the time, the first device created could either be eth0 or eth1 and
then net0 could sometimes point to eth0 and sometimes to eth1 which
doesn't help with analysing the metrics.
So, use Map instead of Vec to provide more clarity on which interface
the metrics actually belongs to.
Also, instead of net0 name the metrics json object as net_$iface_id
e.g. net_eth0 or net_eth1 which should help in better analysis.
Use "net_$iface_id" for the metrics name instead of "net_$tap_name" to
be consistent with the net endpoint "/network-interfaces/{iface_id}"

Note: this is part 6 of 12 commits to add per net device metrics.

Signed-off-by: Sudan Landge <sudanl@amazon.com>