chore(metrics)(net): remove new() for NetDeviceMetrics and proxy
NetDeviceMetrics was moved from logger which needed the new()
because default didn't provide the const implementation that
FirecrackerMetrics::new() needed. But since that is not a requirement
anymore we rely on default to provide 0 initialized NetDeviceMetrics.

NetMetricsSerializeProxy doesn't have any fields and so there is no
need to define new() for it.

Note: this is part 9 of 12 commits to add per net device metrics.

Signed-off-by: Sudan Landge <sudanl@amazon.com>