test: save Firecracker logs if a test fails
It would be useful for troubleshooting to have the Firecracker logs, in
case they can shed some light on the cause of the failure.

This can be done at the microvm_factory fixture teardown, before we
delete each microvm. The only difficult part is getting the test result,
which pytest doesn't provide, but there is a workaround in the docs.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>