fix: stub out set_avail_event call from try_enable_notif harness
Calls to set_avail_event tend to cause memory to grow unboundedly during
verification.  The function writes to the `avail_event` of the virtio
queue, which is not read from by the device. It is only intended to be
used by guest. Therefore, it does not affect any device functionality
(e.g. its only call site, try_enable_notification, will behave
independently of what value was written here). Thus we can stub it out
with a no-op. Note that we have a separate harness for set_avail_event,
to ensure the function itself is sound.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>