test(drive): add support for drive creation on disk
So far, FilesystemFile was only creating drives
in the default /srv location, which is mounted as tmpfs.
Since Qemu vhost-user-blk backend always uses O_DIRECT
when opening drive files, it cannot work with files on tmpfs.

If the path argument is not provided to the FilesystemFile
constructor, the drive file is created with a random name in /tmp.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>