test(ab): artificially increase network latency test cases
Splits the network latency test, which currently does 900 pings from
guest to host into two tests each doing 450 pings. The rationale behind
this is that we have very little network latency test cases (only 2,
which is the number of guest kernels), which means that the A/B-testing
framework does not have enough data to meaninfully correct for outliers
(it has only 2 data points). This change increases the number of data
points it can work with to 4, which should hopefully help with the high
false-positive rate we have been seeing from this test.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>