snapshot: remove Versionize serialization functions
We do not need these any more, because we will never serialize to older
versions of Firecracker snapshots. Also, remove various unit tests that
were checking for failures depending on which version we are serializing
to. This clean up in the tests is not complete. There will be a general
re-structure once we fix the snapshot versioning format.

Signed-off-by: Babis Chalios <bchalios@amazon.es>