test: avoid recompiling examples for every agent
We compile the UFFD and seccomp example programs as a session fixture.
If we run the tests in parallel, there is one session per worker and we
end up downloading and compiling the examples for each worker (worst
case).

Instead, use the same approach as the Firecracker and jailer binaries.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>