fix(test): test_cargo_audit fails on nightly
The test fails when not run as part of PR tests. The issue is that these
tests behave differently when running as part of a PR or independently.
This makes testing and troubleshooting more difficult.

Workaround the issue by changing the directory so the command runs at
the right place.

Fixes: 8297c8f3a5cddc3b709f3ee26db913cf66fbcbb6

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>