feat(virtio/mmio): `MmioTransport` knows about vhost_user
Now `MmioTransport` stores info if the device uses
vhost_user implementation. Based on this info the
interrupt status reported to the guest is either
an updatde interrupt status (for usual virtio devices)
or `VIRTIO_MMIO_INT_VRING`(0x1) for vhost_user devices.
This is done because vhost_user devices can not update
interrupt status variable in the FC.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>