Replace GCD stubs with a function contract
Introduces a function contract for GCD. The contract is checked against
the implementation to ensure that it is satisfied and can also be used
as a stub in other Kani harnesses for increased efficiency.
The effect is a setup very similar to what was done before manually with
stubs and a check for the implementation, but now with only a single
source of truth (the contract).

Signed-off-by: Justus Adam <justbldr@amazon.com>