refactor(test): Add switch to ignore return code in non-PR context
Commit 978b6d89111c ("fix(test): test_spectre_meltdown_checker_on_host
when not running on PRs") changed the spectre/meltdown checker test to
ignore the return code even outside the PR context. To make it clearer
that `ignore_return_code` is only valid in the non-PR context, this
commit renames it to `ignore_return_code_in_nonpr` and adds the same arg
to `git_ab_test_guest_command_if_pr` as well. When ignoring the return
code, we sometimes want to consume the command result on the caller
side, so this commit also makes them to return the result.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>