fix(test): Check return code and ignore inception on m6a/4.14
Commit 978b6d89111c ("fix(test): test_spectre_meltdown_checker_on_host
when not running on PRs") made the test ignore the return code, but
outside the PR context (A/B testing context) we should check the return
code. However, there is an exception right now that the test fails the
inception check on m6a.metal + kernel 4.14, because the upstream kernel
didn't backport the inception mitigation patch to 4.14. To get full
mitigation against inception, we enable IBPB on kernel 4.14 instead of
safe RET and the status can be checked on retbleed sysfs.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>