test: Run fingerprint monitoring as part of CPU template pipeline
The test comparing guest CPU config between the saved and current
fingerprint was part of the security pipeline. Since now it is the only
test in the pipeline and it is related to CPU template feature, this
commit moves the test to the CPU template pipeline and removes the
security pipeline.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>