snapshot: switch to using serde instead of Versionize
We want to simplify the way we handle Firecracker snapshots. Versionize
crate was created as a mechanism to help us keep track multiple versions
for each struct we want to (de)serialize in a Firecracker snapshot. This
allowed us to support forwards & backwards compatibility across multiple
Firecracker versions. In reality, we don't really make use of this
compatibility flexibility. Instead, we keep maintaining multiple
versions for struct fields (even for versions no longer supported),
which increases code and testing complexity.

We will revert to simply using serde for serializing Firecracker state
into a snapshot file. We will also decouple Firecracker from snapshot
versions. The snapshot format will now have an independent version which
starts from v1.0.0.

This change touches many parts of the code base, so we will do it
gradually. In this commit, we change the snapshot crate and the
snapshotting API. We only make the minimum changes needed to the rest of
the codebase to keep Firecracker working. We change `MicrovmState`, the
structure that describes the Firecracker state we serialize in a
snapshot to skip serialization of all its fields (essentially it is now
describing an empty snapshot). In subsequent commits, we will start
re-adding fields as we adapt the relevant structures from Versionize to
serde.

Integration tests will not be working for this commit.

Signed-off-by: Babis Chalios <bchalios@amazon.es>