snapshot: serialize VmState and VcpuState
Derive Serialize and Deserialize and remove Versionize support. Fix
these two structs together because both depend on kvm-bindings
implementing (De)serialize for its types, so updating to the
kvm-bindings to the version that introduces serde breaks both of these
types.

While implementing serde for VcpuState also remove the `msrs` field.
According to Versionize this was end-of-life and we are anyway breaking
compatibility with previous snapshot versions.

Signed-off-by: Babis Chalios <bchalios@amazon.es>