fix(test): microVM config after restore
We were skipping the serialization of the "boot_args" field of the
"boot-source" objectm, if kernel boot arguments were not supported. Now,
we return None. This was done so that we have consistent snapshot
formats, now that we use serde to serialize microVM state in a snapshot
file.

Amend this test to expect '"boot_args": None' in the "boot-source"
object.

Signed-off-by: Babis Chalios <bchalios@amazon.es>