fix(test): remove tests that are not relevant any more
With the new snapshot format we break compatibility with all previous
snapshot format versions. This commits removes all the tests that were
checking functional correctness and performance of loading older
snapshots in the current Firecracker binary.

In the future, it might be that two Firecracker binaries support the
same snapshot versions. When that happens we should thing of a proper
way to test that this works, but the current tests are not fit for
purpose.

Signed-off-by: Babis Chalios <bchalios@amazon.es>