release: Copy custom CPU template JSON files to release directory
The custom CPU template JSON files are useful for users to migrate from
static to custom CPU templates smoothly.

Note that this does not release them in GitHub releases, since they are
already available on GitHub without any action (like building a binary).
When we get a reasonable request from GitHub release consumers, we can
release them in it.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>