test(vhost-user-blk): switch to crosvm backend in functional tests
Unfortunately this change cannot be split into
multiple commits due to high level of entanglement
of the changes.

Reason for the change: the Qemu backend uses direct IO
when working with the disk file which makes it
scale very poorly.  That results in intermittent
test failures, because our functional integration tests
run in parallel, and our ssh connection timeouts
expire if the microVM does not boot in time.

Switching from the Qemu to crosvm backend comes with
the challenge that crosvm locks the disk file exclusively
when opening it as rw (which makes sense).  We have
to make copies of the disk files before we give
them to crosvm if we don't configure crosvm as readonly.

One of the ripples of the above is an existing race
condition in the partuuid_and_disk_path_host fixture
over the disk file kicks in.  We are removing this
fixture altogether because now we use per microVM
copies of the disks.

Finally, this change explicitly configures
performance tests to switch to the Qemu backend
to preserve the status quo.  We may switch them
to the crosvm backend too at a later stage.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>