fix(drive): made FileEngineType optional
Updated `BlockDeviceConfig` to have `file_engine_type`
as an optional filed, instead of a filed with default value.
This is done, because for vhost-user-block this field
is not used, but it would still be have some value when user
would request the block config.
With this change config for vhost-user-block will return
`None` for `file_engine_type`.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>