test: Add a test for taking diff snapshots on top of full snapshots
The test first crates a full snapshot, and then takes a diff snapshot
with the same memory file, to merge it. We take both snapshots as "diff"
snapshots, as taking a full snapshot does not clear the dirty pages
tracked (and in this test, we want to exercise the path where the second
diff snapshot only writes a few pages).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>
Co-authored-by: Ives van Hoorne <ives@codesandbox.io>