fix: intermittent vsock timeout in snapshot tests
Instead of creating a new socat process on every guest resume,
create socat process in the 1st guest created so that the snap shotted
guests just use the same process through.
Suspend and resume socat process before taking snapshot and after
resuming from snapshot respectively to make sure the process is in
the right state for the snapshot.
The socket connection is reset by Firecracker while creating a
snapshot so it seems to leave socat in a weird state and without
stop/cont 'ng socat before/after snapshot we end up in a vsock timeout
when host tries to connect to guest in _vsock_connect_to_guest().

Signed-off-by: Sudan Landge <sudanl@amazon.com>