fix: Move "should run tests" from buildkite to github action
Our internal check to alarm us if a post-PR A/B-test fails does not fire
if a "bailed out" run (e.g. one where we decide not to run A/B-tests
because we only modified, say, .md files) happens while a previous
A/B-test is still running. This is because it only looks at the most
recent buildkite run.

Therefore, move the check whether to run an A/B-test at all from
buildkite to the GitHub action that orchestrates the runs, so that "bail
outs" happen outside of buildkite.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>