chore(seccomp): remove unused syscalls from vmm thread
This removes the following syscalls
from the VMM thread seccomp filters:
- clock_gettime
- exit
- mremap
- rt_sigaction
- mmap(MAP_SHARED)
- ioctl(TCGETS)
- ioctl(TCSETS)

We have a datapoint that at some point after adding
extra syscalls to a filter, we observe performance
degradation on certain workloads (specifically
host to guest TCP throughput on Ice Lake).

This change is supposed to address that problem
and also improve security by allowing only
the syscalls that are actually used.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>