test(metrics): Add new test to collect Firecracker vhost-user metrics
Add a test to boot a VM with vhost-user as a scratch device.
Resize the vhost-user scratch drive to have config change notifications
and collect vhost-user metrics emitted by Firecracker for this test.
Export the metrics to CloudWatch.
We use a new file in performance directory for this test becuase:
1. Tests doesn't need to be ABfied so can't be part of test_block_ab
   and the Performance pipeline.
2. By running in nightly pipeline keeping the test in functional
   directory, there will be lot of other tests running in parallel
   resulting in noisy metrics.
3. Nightly pipeline runs performance tests in a decidated BK agent,
   with nothing else running on the host, and with no parallelism so
   it makes sense to move to performance directory but keep it out of
   AB tests.

Signed-off-by: Sudan Landge <sudanl@amazon.com>