fix(test): don't try to kill after CtrlAltDel
In test test_send_ctrl_alt_del we send a CTRL+ALT+DEL to the microVM,
which, in x86, makes the microVM to shutdown. Then we send a signal to
the Firecracker process with `os.kill(firecracker_pid, 0)` and wait for
it to fail. This works but logs an error in the test logs which can be
confusing.

Instead we can call os.waitpid() which waits for the Firecracker process
to exit and returns immediately if the process has already exited.

Signed-off-by: Babis Chalios <bchalios@amazon.es>