fix(test): do not pass network config to CPU template helper
CPU template helper tool uses a microVM configuration we pass to it to
launch a VM and get it's CPU config. Our integration test that drives
that needs to configure the microVM with a network interface so that it
can SSH into it and run commands. However, when we pass to helper tool
a configuration that includes networking, we run into a race-condition
with creating/configuring a TAP device.

This commit passes a microVM configuration without a network interface
to the helper tool to avoid the race.

Signed-off-by: Babis Chalios <bchalios@amazon.es>