fix: Make test_cmp_full_and_first_diff_mem not cmp file to itself
The Microvm.snapshot_*() functions by default always write the snapshot
to the same file system location. This means that full_snapshot.mem and
diff_snapshot.mem point to the same path, making the final assert in the
test trivially true (as a file is the same as itself - the diff snapshot
will have overwritten the full snapshot). The fix is to use a different
path for the full snapshot.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>