test: Overwriting snapshot from which microvm is loaded
This commit adds a test that has a microvm overwrite the snapshot file
from which it was loaded with a new snapshot (e.g. a sort of "refresh"
operation on the snapshot to write any modifications done since loading
back to disk). Prior to this patch series, this caused both the snapshot
and the microvm to become corrupted, as the zeroing done by the
truncate/set_len pair would be reflected in the mmap'd memory region.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>