test: Parameterize tests that use guest features by guest kernel
For historic reasons, a lot of our integration tests uses the
test_microvm_with_api fixture, with caused them to only use a 5.10 guest
kernel. For tests that meaningfully interact with the guest kernel, we
instead want to test all supported guest kernels. The uvm_plain_any
fixture does exactly this, so use it.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>