fix(jailer): account time spent in the jailer more accurately
After da92bf6d, the CPU time we got from get_time_us was much shorter,
and that caused an underflow when calculating the jailer_cpu_time_us.
That resulted in Firecracker getting called with a huge value for
`--parent-cpu-time`.

Before this fix:

    /firecracker [...] --parent-cpu-time-us 18446744073709551368

After this fix:

    /firecracker [...] --parent-cpu-time-us 4375

The new time is higher than before da92bf6d, because a) that commit
introduced a fork, which carries some overhead, and b) we now account
for CPU time during forks.

Fixes: da92bf6d4afa678565dcf9df2e4e00d2fa273d96

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>