feat(jailer): honor solitary --parent-cgroup option on cgroupsv2
If we specify `--parent-cgroup` without any `--cgroup` option, the
jailer doesn't do anything, though a user specifying it could reasonably
expect it to move the process under that cgroup.

We could just error in that situation, and expect something else to
launch the jailer process in a cgroup, but we risk breaking anybody that
is already using it that way.

Instead, we move the process to the `--parent-cgroup` since it's the
most intuitive, although the specified `--parent-cgroup` is not a parent
in that case.

Link: https://github.com/firecracker-microvm/firecracker/issues/4287

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>