perf: Use smallvec::SmallVec for IoVecBuffer[Mut]
IoVecBuffer[Mut] use a Vec of iovecs to hold all virtio descriptors of a
descriptor chain. This incurs an allocation costs that can be seen in
vsock throughput data, if the vsock device code is switched to using
IoVecBuffer[Mut] to handle its descriptor chains. However, we know that
heuristically most descriptor chains are short (e.g. the linux guest
uses 3 descriptors for virtio-block requests, at most 2 for virtio-vsock
requests). Therefore, we can optimize for this common case by using a
SmallVec, which stores a small number of iovecs (up to 4 in our case) on
the stack, to avoid the allocation cost.

Fixes #4203

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>