Fixup vsock unit tests to work with iovec-based packet handling
IoVecBuffer[Mut] statically guarantee that a read-only descriptor chain
cannot be written to, and a write-only descriptor chain cannot be read
from. Since RX descriptors are write-only and TX descriptors are
read-only, this causes some problems in the vsock unit test code, which
has to first emulate the guest putting packets into the TX queue (e.g.
write to a read-only queue) and also reading the data that the device
wrote into the RX queue to validate its correctness (which is a
write-only queue). The unit tests however do not differentiate well
between which queue a packet belogns to, using one packet that was
initialized from the RX queue to both mock RX and TX operations. This
commit fixes this somewhat, by differentiating between RX and TX
packets. It does not address the separation between guest- and host-view
of the respective queues, and instead uses a trick: Both queues point
toward the same memory region, which is then interpreted as RX or TX
memory depending on whether the test wants to mock a guest operation, or
validate a device operation.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>