chore: Remove unused VsockError::BufDescMissing
With the vsock code no longer asserting that every vsock packet has two
packets, this error variant was no longer constructed. It's only use was
in a unittest elsewhere, that tested that a vsock error was properly
propagated (however, this test does not require this specific error, so
we can reuse a different variant as the dummy value).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>