test: A/B-ify boottime test
Run the boottime test as an A/B-test post-PR so that we note changes to
this key metrics other than "are we below 150ms?".

Also incorporates the changes from #4316 so that we can run the
short-running boottime and memory overhead tests on the same .metal, to
avoid spawning yet another yet of 15 .metals after every PR merge.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>
Co-authored-by: Jonathan Woollett-Light <jcawl@amazon.co.uk>