refractor: flesh out semantics of IoVecBuffer[Mut]::read/write_at
Before this patch series, IoVecBuffer[Mut] only ever had to deal with u8
slices, which inherently have an "all or nothing" semanticfor Read and
Write. Thus the semantics of read_at and write_at were "return None if
the given offset is too large this for IoVecBuffer[Mut], otherwise copy
either buf.len() or iovec.len() - offset bytes, whichever is less". This
commit changes the second part of this behavior to "copy either
buf.len() bytes, or fail", which is how these functions were used in
praxis. It also brings the semantics more in line with read and write
functions offered elsewhere in the standard library. For now, it still
only operates on u8-slices, but should the future need arise to
generalize to Read/WriteVolatile, all the pieces are there.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>