Add option to run A/B-test on two arbitrary test-report.json files
This allows comparisons of two test runs that were not orchestrated by
the A/B-testing framework, for example if the two runs were done on
different instances. Use this if the data you want to compare is from
anything other than "two firecracker revisions tested on the same
machine inside the same docker image".

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>