fix(metrics): Add more validation to check missing metrics
validate_missing_metrics is actually just validating the jsonschema
library and not any Firecracker property itself which is fine since
it is kind of a reassurance that it is working as we expect it to.
However, atm it only checks if jsonschema throws an exception on
the missing metrics and it should in fact also check if there was an
exception raised at all. If for some reason jsonschema doesn't raise
an exception then the test passes and the purpose of this function is
defeated.
So, raise exception if jsonschema doesn't raise one.

Signed-off-by: Sudan Landge <sudanl@amazon.com>