vmm: handle uniformly GSIs and memory of a guest
Right now we handle guest interrupt lines (GSIs) and physical memory
inside the MMIO device manager.

With ACPI we will add a new device type for which we will need to
allocate GSIs and memory. With the current way we do things, this means
that we would need an extra pair of GSI and memory allocators.

This commit introduces a resource manager at the VMM level for handling
GSIs and guest physical memory allocations. Each device manager will get
a reference to this object for handling resources globally.

Signed-off-by: Babis Chalios <bchalios@amazon.es>