fix: bump rust-vmm dependencies
Update dependencies for various rust-vmm crates which made new releases
to adapt to changes in vmm-sys-util crate. The vmm-sys-util changes were
triggered due to a security vulnerability discovered in vmm-sys-util
(Github Security Advisory id: GHSA-875g-mfp6-g7f9), affecting the serde
implementation for the FamStructWrapper types.

Firecracker itself is not affected by this vulnerability since we do not
consume the serde feature of the FamStructWrapper type, but we will in
the future. So, update the dependencies now to prepare for this and also
avoid having `cargo audit` complaining to us.

The most notable change here, is that we need to update our fork of
kvm-bindings to bring it in-par with the upstream version.

Signed-off-by: Babis Chalios <bchalios@amazon.es>
Signed-off-by: Sudan Landge <sudanl@amazon.com>