fix(test): correct kernel check in test_shut_down
PR #4353 upgraded versions of various rust-vmm crates that got patched
for a security vulnerability found in vmm-sys-util (not affecting
Firecracker). This included a new version of kvm-bindings which includes
bindings from a newer kernel. In this, `struct system_event`, which is
the type that KVM passes over to us during a KVM_EXIT_SYSTEM_EVENT exit,
changed from being a `u64` to being a `&[u64]`. The change itself,
landed in kernel 5.18.

This changed, altered Firecracker logging during shutdown (in some
cases), so #4353 PR made the necessary changes to fix the tests that
relied on the log entry. The fix boils down to "if we are on 6.1 check
for the new type of message". That caused our testing on latest Ubuntu
to fail, because now it uses a 6.2 kernel.

Fix the test by correctly checking if we are running on a kernel >= 5.18
(the kernel version that introduced the change in the KVM type).

Signed-off-by: Babis Chalios <bchalios@amazon.es>