refactor: Hoist `update_vm_config` call to begin of restore path
Ensure that the vm_config field of the `VmResources` struct is
initialized from snapshot data as early as possible. This means that
further steps in the snapshot restore procedure can rely on the
vm_resources object for vm configuration information, instead of needing
to access the microvm state (or passing various parameters down to ever
snapshot restore function aslong with the uninitialized `VmResources`
struct, as was done for dirty page tracking). This change also ensures
that we fail early if the snapshot is malformed in some way, as opposed
to failing after various KVM resources have already (potentially
incorrectly) been initialized.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>