test: Add test that booting with hugetlbfs memory works
Attempts to boot a microvm with guest memory backed by 2MB hugetlbfs
pages. Adjusts the test infrastructure to allocate 2MB pages prior to
test run (failing if it cannot do so). For now, we rely on no other
process on the host trying to use hugetlbfs.

The test is skipped on 4.14 because hugetlbfs support for sealable
memfds was only added in 4.16.

We put the test as a performance test to ensure it runs on ag=1 agents,
to avoid problems with different agents on the same metal concurrently
modifying the hugetlbfs pool.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>