acpi: create acpi-tables crate
Add a lib crate for handling ACPI tables. The crate exports types
representing the various tables we handle and methods for creating
them, modifying them and writing them in guest memory.

The crate is based on the cloudhypervisor implementation:
https://github.com/rust-vmm/acpi_tables. We reuse the logic for creating
AML bytecode but we diverge from the way they create and manipulate the
tables. They rely mainly on the SdtHeader type and work with byte
offsets, whereas we define concrete types for each table and use those.

Signed-off-by: Babis Chalios <bchalios@amazon.es>