fix: Add check for known vulnerabilities in guest
There is a REPTAR exception reported on m6i.metal for
guest kernel versions 4.14/5.10/6.1 when spectre-meltdown-checker.sh
script is run inside the guest from below the tests:
test_spectre_meltdown_checker_on_guest and
test_spectre_meltdown_checker_on_restored_guest
The same script when run on host doesn't report the
exception which means the instances are actually not vulnerable to
REPTAR.
The only reason why the script cannot determine if the guest
is vulnerable or not is because Firecracker does not expose
the microcode version to the guest.

The check is check_CVE_2023_23583_linux in spectre-meltdown-checker.sh

Since we have a test on host and the exception in guest is not valid,
we add a check to ignore this exception.

There could be more such exceptions added to
check_vulnerabilities_on_guest() in the future but generalising this
function based on instance kernel versions etc. without more data
looked like overkill right now so,
the function handles the exception only for REPTAR and m6i.

Signed-off-by: Sudan Landge <sudanl@amazon.com>