feat(uffd): new Runtime
The new `Runtime` struct can contain multiple
uffd objects received from the UDS socket during
its runtime.

This is done by polling the socket and uffd
file descriptors. When the socket event is
polled, a new uffd object is retrieved. When
the uffd fd is polled, a page fault for
corresponding uffd is handled.

The behaviour for page faults can be configured
with a parameter passed to the `Runtime::run`.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>