perf: Change virtio event_handlers to use Events::with_data()
Using Events::with_data when registering events for block/network/vsock
devices and associating a u32 const data with each event. Then, in the
process function, instead of using if to see what process_* function to
call, we can do a match on consecutive u32 values which is going to be
optimized by the compiler into a jump table.

Signed-off-by: David Nguyen <nguydavi@amazon.com>