fix: reset dirty log on full snapshot path
When a diff snapshot it taken, it always contains all pages that were
dirtied since either instance start, or since the last diff snapshot was
taken, whichever is more recent. This is not what we document it to do:

> the diff consists of the memory pages which have been dirtied since
> the last snapshot creation or since the creation of the microVM,
> whichever of these events was the most recent.

Here "last snapshot creation" includes full snapshots, not just diff
snapshots (and that makes sense, if I take a diff snapshot after a full
snapshot, I expect the diff snapshot to be a diff compared to the full
snapshot).

Signed-off-by: Jack Thomson <jackabt@amazon.com>
Co-authored-by: Roman Kovtyukh <HelloDearGrandma@gmail.com>
Co-authored-by: Muskaan Singla <msinglaa@amazon.com>