acpi: drop MPTable support
We are now using ACPI MADT table to pass to the guest info about
interrupt handlers and CPUs. Drop support for MPTable. Also, re-organize
the memory layout. Use EBDA to store ACPI tables and mark the rest of
the region up until the beginning of the high memory as free for use.

Signed-off-by: Babis Chalios <bchalios@amazon.es>