acpi: define x86 devices in ACPI DSDT table
Use AML bytecode in DSDT ACPI table to define VirtIO devices as well as
passing them as kernel command line arguments. Also, pass AML for legacy
devices, as the kernel doesn't pick up the console or i8042 unless we
do.

For the time being, we are keeping the old way of booting, i.e. MPTable
plus VirtIO devices via command line arguments to allow for some sort of
backwards compatibility with kernels that are not built with ACPI
support.

Signed-off-by: Babis Chalios <bchalios@amazon.es>