ci(perf/net_latency): increase absolute delta threshold to 10us
In some instances we see AB tests failing with very small deltas:

    A/B-testing shows a change of 1.75μs, or 5.46%, [...]

In Graviton instances the 5% default threshold very narrow since the
absolutes values are smaller, and we get false positives like the one
above, that are mostly due to noise.

Set the absolute delta threshold for network latency test to 10us, which
is chosen arbitrarily but should reduce the number of false positives.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>