chore: Upload FCMetrics with reduced dimensions
FC metrics were uploaded to CW with the same dimensions
as the test to make it easy to map the metrics with the test
in CW. However, each test has a huge number of dimensions which
when multiplied with the number of FC metrics we upload add upto
too many datapoints stored in CW.
This makes CW slow hard to monitor the dashboard and also has
a significant increase in billing.
So limit the dimensions to (cpu-host_version-guest_version) so
that we can monitor the dashboard properly and reduce the billing.
To map the datapoints with the test that emitted them we'll have
to use the datapoint timestamp with some CW query.

Signed-off-by: Sudan Landge <sudanl@amazon.com>