fix: test_sec_audit: ignore cargo audit output
Since we run this test twice in case of a PR, cargo audit locks the
database and one of the processes may print something (even with -q
specified) in stdout (‽). I believe this should be printed in stderr,
but we can workaround it here easily.

    warning: directory /usr/local/rust/advisory-db is locked
    {...}

Use `grep` to filter out the extraneous output.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>