test: Add metric tracking the number of EPT_VIOLATIONS post restore
An EPT_VIOLATION kvm_exit happens when the MMU's extended page tables
are missing an entry for some guest physical address (e.g. some page is
present in the guest page tables, but KVM has not yet set up a mapping
of guest-physical->host-physical address for it). This happens after
snapshot restore even if a page is faulted in via UFFD, as fauling in
via UFFD only maps the page into host userspace (e.g. Firecracker), but
does not set up the EPT entries.

We track the number of EPT_VIOLATIONS post restore when using UFFD for
both 4K and 2M pages, as we expect their number to be significantly
lower when using huge pages.

We use a special UFFD handler that faults in the entire guest memory
ahead of time, as otherwise we just track normal page faults.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>